Update to use sql server

+++ InventoryAPI/Models/Image.cs
@@ -7,9 +7,9 @@ namespace InventoryAPI.Models
 {
     public partial class Image
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string Description { get; set; }
         public string ImageName { get; set; }
-        public long? ItemId { get; set; }
+        public int? ItemId { get; set; }
     }
 }
