+++ InventoryAPI/ViewModels/OrderViewModel.cs
@@ -10,9 +10,9 @@ namespace InventoryAPI.ViewModels
         public int? ItemId { get; set; }
         public string Description { get; set; }
         public int? Quantity { get; set; }
-        public int? Price { get; set; }
-        public int? Tax { get; set; }
-        public int? Total { get; set; }
+        public decimal? Price { get; set; }
+        public decimal? Tax { get; set; }
+        public decimal? Total { get; set; }
 
         public Order ToEntity()
         {
